Add InSPIRation Pass, Add Temporary Hacks, Add Working Directory with .sh and Makefile for XOCC
A few workarounds to get things moving in the right direction initially:
1) SemaSYCL.cpp: Currently it doesn't mangle the name of kernels as the 
unusual characters like $_ pose problems for XOCC when used as a kernel 
name
2) queue_impl.hpp: I currently comment out the CL 2.0 way of creating a 
command queue as the way the runtime currently checks for the CL_VERSION 
doesn't appear to be correct and at the very least doesn't work in our 
situation
3) program_manager.cpp/.h: Not neccessarily a workaround more a WIP 
(both by us and Intel by the looks of it), but the current 
implementation of the runtimes program_manager seems a bit specific to 
SPIRV so I added in a getBinarySource call for the time being and 
removed an unused variable (it may be possible just to leverage the 
original getSpirvSource call but it assumes .spv as the file extension 
and requires an environment variable to be set at the moment)

This commit adds a very barebones InSPIRation pass as it no longer needs 
to do 99% of what it used to do, the pass mostly gives arguments to 
kernels generic names and fixes the metadata to be spir-df compliant.

This commit also adds a new top level folder working-directory which is 
there to store tests I'm working on and the Makefile/Shell script for 
generating XOCC and a setup Shell script for environment variables that 
are used in the Makefile and by the SYCL compiler (perhaps needs 
renaming from set_intel_sycl.sh). It also stores the original diff of 
the sycl patch.