Initial implementation of SYCL XOCC one command offloading
To compile a kernel for xocc setup all the relevant environment 
variables, as shown in set_xilinx.sh (importantly XCL_EMULATION_MODE 
toggles what the shell file will compile to). Then compiling a cpp file 
with several kernels should be as simple as:

clang++ -std=c++11 -fsycl-xocc-device single_task_vector.cpp -o 
single_task_vector -lsycl -lOpenCL

Remember that you'll need a Xilinx scheduler in your project to 
guarantee selection of the right OCL runtime or it may choose the 
incorrect runtime and try load the binary to the incorrect 
implementation. Some basic examples are in working-directory that 
contain schedulers for pocl, iocl and xocl.

XRT (Xilinx Runtime, OCL runtime amongst other things) will also die if 
you run a binary that was compiled for a target with the wrong 
XCL_EMULATION_MODE enviornment setting. As XRT seems to currently rely 
on XCL_EMULATION_MODE to choose some internal paths for the input 
binary.

Changes:

New compiler flag -fsycl-xocc-device, up for debate whether a flag is 
the best direction to take this, but it makes putting a quick 
implementation together easier for the moment. The flag is used to turn 
off kernel name mangling and to tell the SYCL Toolchain to use the XOCC 
Assembler/Linker as well as to tell the Clang Toolchain -S (output of: 
ty_PT_ASM) stage to output llvm ir rather than the normal output 
(ToolChains/Clang.cpp line 3611, important change as its what feeds our 
assembler llvm ir instead of something unpalatable). This is in the same 
vein as the Intel modification to force the Clang Toolchains Assembler 
stage to output spirv based on their flags (same file line: 3575~). It 
would be possible to consolidate these changes if we combined our 
Assembler and Linker into one Linker phase.

Added KernelNameGen pass which is a very simple pass that just finds the 
kernel names in an ir file and prints it to a text file. Can be expanded 
on as we wish to harvest other data to pass to XOCC if required. Or 
merged into InSPIRation.

New sycl-xocc script that takes a data file generated by KernelNameGen 
pass and invokes either the XOCC linker stage or compile stage based on 
arguments passed to it. The directory/location this should exist and 
where abouts it gets built too are till open for input, much like the 
majority of this implementation. This script should also be made more 
modular and generic as we add to it.

Modified the SYCL driver to include an XOCCLinker/XOCCAssembler which 
basically handles the invocation of the sycl-xocc shell script. Again a 
lot is up for debate on this as its an initial rough implementation. For 
example there isn't really a need to have a seperate 
XOCCLinker/XOCCAssmebler, could in theory merge into one Linker phase 
which would simplify a lot. The XOCC components may also be better off 
in their own Toolchain file rather than being nested inside the 
SYCL.cpp/h Toolchain. It's also possible to remove the llvm-opt and 
llvm-link construct commands from the Assembler phase and put them into 
the shell file if we wish. This would further tidy things up and keep 
more of our implementation details external to Clang.

Some minor modifications to program_manager.cpp, get_binary and 
get_spirv are essentially the same call at this point, so maybe I should 
revert it to just use get_spirv even though it's technically not spirv. 
Makes merging upstream intel-sycl implementation less of a hassle, sort 
of just being pedantic about the name at this point.

set_xilinx.sh in the working-directory, which just shows how I set 
things up right now. It may need to be removed or modified if I set this 
to public as it has internal paths.

That's the gist of the changes in this commit, a lot is open to 
modification if this isn't the ideal direction.