Initial modifications for further built-in support
This commit includes changes to the inSPIRation pass that remove the 
mangled namespace from functions inside the cl::spirv:: namespace and 
modifies their ZN_ encoding. The goal is to modify OpenCL/SPIR built-ins 
inside the cl::spirv namespace to the appropriate spir mangling so they 
can be appropriately linked with a spir header that implements these 
built-ins.

NOTE: The InSPIRation pass is only run when compiling to XOCC and at the 
moment it SHOULD NOT be applied to anything where the intent is to 
output SPIRV/CL++. As the InSPIRation pass currently doesn't care what 
built-ins get ran through it, it will alter the call. In reality the 
SPIR built-ins should be in a strict cl::spir namespace to avoid this.

With this the currently implemented math functions should work (except 
for mad which is broken perhaps on the HLS side rather than by us, need 
to discuss with Alexandre) for C++ built-in types like float, int. Have 
not tested with vectors yet as Intel are still working on a fix for the 
math functions to be useable.

I added a new define __SYCL_SPIR_DEVICE__ that needs to be defined when 
compiling for a SPIR target (either directly to clang via -D or inside 
the source file being compiled). This is important as the get_* 
functions of OpenCL are walled behind this right now, the math functions 
will currently work however.