Adding simple Xilinx vendor extensions from triSYCL, some new tests and upping SYCL runtime std to 20
I added a CL/sycl/xilinx folder for xilinx extensions and brought across 
the partition_array, ssdm_inst and opt_decorate_func files from triSYCL. 
This allows the use of the partition_array and pipeline extensions.

Updated the edge_detection.cpp test to be functional for Xilinx and 
Intel based on a XILINX define, the difference is that when compiled 
with the XILINX define it makes use of some of the intrinsics. I also 
use some more C++ functionallity from the STD library using a work 
around that defines standard library components only when compiling for 
the host.

This commit raises the SYCL runtime compilation standard from 11 to 20. 
The C++ stubs of the standard library still need to be brought up to 
date with C++ 20 for host code to actually be compileable with 2a. At 
the moment it'll only go up to 17 without issues with the stubs.

Also removal of some temporary files I accidentally pushed in a prior 
commit related to the edge_detection example.

Addition of some small tests that test various things:
constexpr_correct.cpp - check if constexpr works as expected on device
no_device_printf.cpp  - making sure printf doesn't work on the device
math_host_conflict.cpp - wanted to check if any host conflicts with 
current math functions, not a particularly complete test case
median_filter.cpp - this is the beginning of a port of the SDAccel 
example of the same name, it's a major wip still.