Adding the Hashing of Kernel names and reqd-work-group-size as a property
Adding kernel name hashing to the run-time in program_manager and the 
llvm-pass, uses boost hashing of a sha1 uuid.

Removed the SemaSYCL.cpp work around that stops mangling of kernel names 
when fsycl-xocc-device is given as an option.

Fixed the edge_detection test as it was breaking in HW synthesis as the 
xilinx pipeline wasn't in a great spot, the loop nest needs to be more 
optimized for it to go on the outer loop. Also removed some more garbage 
files I accidentally included.

Added nested namespaces to the Xilinx vendor files.

Added reqd-work-group-size property to xilinx vendor and added 
functionallity to the inSPIRation pass to search for this and apply the 
correct metadata to the function. Added a very small new test to check 
the output IR for it called reqd_work_group_size.cpp.