Addition of integration_header_check.cpp test to issue_related
This test checks various ways to name kernels to see if they:
  1) Compile
  2) Execute correctly

The final two test cases will not work correctly without a future commit 
that alters the compiler, the ability to add properties (like 
reqd_work_group_size) to kernel names is reliant on the change. As the 
host won't compile properly without it, as the integrated header will be 
generated incorrectly (up for debate perhaps with Intel when I put in a 
pull request).

This test is fairly identical to the kernel_name_check.cpp test in the 
SYCL regression tests. I made it without realizing this unfortunately.