Reverting early changes to program_manager.cpp that were causing some issues
Some changes to my CL driver made searching for the il_ spirv extension 
for intel devices invalid which caused some trouble when doing two step 
compilation. Reverting the change back to what it used to be just to 
make life easier. The only real difference is that if you want to two 
step compile a spir kernel you going to have to prefix it with .spv even 
though it isn't spirv.