reqd_work_group_size property run-time experiment
The data stored in the property can now be accessed in the runtime where 
you have access to the string form of the kernel name. You can retrieve 
the data using the get_reqd_work_group_size function added to the 
kernel_properties.hpp.

This method isn't an ideal way of doing this, but it works for the time 
being, it uses the same method as in the InSPIRation pass and demangles 
the mangled kernel name then regex's for the data. A more compile time 
friendly method would be better in the future. Perhaps some kind of 
property manager.

The data retrieved from the property is currently used to set the local 
work size inside command.cpp for parallel_for's that have nd_range 
specifying their local work group size and to set the single_task. It is 
debateble if this has any use though, as there is no way to get the 
local work group size in this type of parallel_for. The main goal was to 
fix single_task encountering a run-time error when reqd-work-group-size 
was specified for it. This was occuring because clEnqueue was being 
given a nullptr all the time. However, it can be fixed without querying 
for properties. Just check if the command is a single_task, if it is 
specify the local work group size to be 1,1,1!

edge_detection test was tweaked to use reqd_work_group_size.

New reqd_work_group_size test made to check some basic use cases work 
and do not work where relevant.