Alter CodeGenSYCL tests to pass again
Fixed some CodeGenSYCL tests related to the SYCL integration header as
we currently alter it a bit within SemaSYCL and there was some recent 
additions
to these tests. There is a pull request in the works at the moment so 
we'll see
if we have to revert it or not soon enough.