Driver ToolChain fix and script fix
The XOCC ToolChain was looking in the wrong location for the SPIR 
libraries. A future addition should be to have the option to pass the 
directory down to the driver to make it more robust.

I added the sycl-xocc script as a dependency to SYCL, so when SYCL is 
built the script should get "built" as well and placed in the bin 
directory.

I added the ability to specify the platform to the script using the 
XILINX_PLATFORM environment variable, similar to the driver comment 
above, an option to pass this down to the SYCL XOCC ToolChain would be a 
good idea.