Builtin Maths WIP
This should fix the math functions in the current math_mangle.cpp test 
(and probably a lot more) but additional tests on the now rather large 
range of math functions need to be done. The change was required due to 
some alterations in the reflower/runtime to support newer math 
functions.