Change device compiler standard to equal passed in standard, this seems to allow C++2a compilation
This means the device side compilation will always use the same standard 
as the host when given -std=C++xx.

An interesting side effect of this is that -std=c++2a will now work on 
all the tests for both intel and xilinx compilation, it appears the 
problem was not with the stubs and perhaps with the device compiler and 
host compilers C++ versions being disjointed. Need to keep an eye on 
this, but C++ 2a seems useable for the moment and should be prefered to 
find issues.