Adding internal compiler definitions for -fsycl-xocc-device to clean up command line invocation
__SYCL_XILINX_ONLY__ is now a default define when compiling with the 
-fsycl-xocc-device flag, it is enabled for host and device compilation 
for the moment. Used to optionally turn things off and on for Xilinx 
SYCL compilation in the runtime (xilinx extensions) and in some tests.

__SYCL_SPIR_DEVICE__ defined by the compiler when -fsycl-xocc-device is 
given. ONLY active on the device, this shouldn't be used outside of the 
runtime. It's mostly for flipping the switch to swap in SPIR intrinsics 
for the moment.

Overall the addition of these in the compiler has allowed the compiler 
invocation for xilinx fpga's to become cleaner for the end user and 
allows some simplification and cleaning of tests.