Fixes exception generation for device side code
Ran into an interesting error that reared it's head when trying to get 
xilinx::pipeline to run on intel devices without any nasty user 
requirements. The noexcept specificer on pipeline was causing exception 
related code to be generated on the device, choking llvm-spirv. 
Surprisngly xocc seemed fine with this, but I remember exception related 
things being a problem on triSYCL so this fix helps everyone as 
exceptions shouldn't exist in device code (at least for OpenCL).

Just turns off exceptions for SYCL device compilation, issue opened on 
the main SYCL repository by Intel and agreed this is a bug and can be 
fixed by the method in this commit. Issue #77 noexcept specifier device 
side ICE in LLVM-SPIRV.