[SYCL] Removing requirement to give -fsycl-xocc-device as command line argument
Unfortunately, while the user is no longer required to specify a command
line argument to the user, the compiler still needs a new Option for
information to be pushed from the Driver onto both the host and device
for now. If it's just device related, it would be feasible to check the
Target triple. Sadly the host invocation doesn't contain any knowledge
of the offloaded triple from what I can tell (aux-triple would be nice
to use, but it seems this is largelly used to pass host triple
information to device offloading passes and not vice-versa).

Changed -fsycl-xocc-device -> -fsycl-xocc, as it currently enforces both
host and device related tweaks (mainly forces the host definition of
__SYCL_XILINX_ONLY__, which I think should be refactored out long term
for a runtime solution if feasible).

Moved option -fsycl-xocc from Options.td to CC1Options.td, as it's more
"Clang" specific rather than part of the driver frontend at the moment
(shouldn't really be read by the driver/isn't at the moment). Perhaps, a
bit pedantic...

Adding a test to check internal compiler defines that should print some
messages based on some defines that should be generated by the compiler.