[SYCL][Runtime] Enforcing Default DDR0 Buffer Assignment In Runtime
This is a restriction for newer versions of hw_emu with XRT (2018.3+), 
we must always specify the DDR Bank a buffer is assigned to on a 
Kernel/CU basis.

Perhaps this restriction will not be permanent, but either way a future 
addition would be to have some way to specify in SYCL a DDR Bank 
assignment for a buffer to a specific kernel (accessor makes more 
sense?), I've added a larger \todo relating to this in buffer_impl.hpp.