[SYCL][sycl-xocc] Opt out of Dataflow checks when compiling kernels
This check was causing some problems in hw_emu for newer versions of 
SDx, I couldn't compile what is essentially an empty kernel. The error 
diagnostics seem a little broken at the moment and are not particularly 
useful for tracking this issue down. I have left a todo to come back to 
this in the future.