[SYCL] Adding further important comments and altering the SYCL CMake build file
Removal of -D__SYCL_XILINX_ONLY__ from the build sysytem as its not 
required at the moment and an alternative solution may be better in the 
eventuallity its required.

Added some comments relating to my thoughts on how to handle XRT CL 
extensions in the future in a manner that should be friendly to our 
build system and non-xilinx devices.