[SYCL][Build][Runtime] DDR Buffer Default Assign Fix and application to SCHEDULER_20
This patch:
* Fixes the default DDR Bank assignment for SCHEDULER_10
* Adds the default DDR Bank assingment for SCHEDULER_20
* Fixes CL_INVALID_VALUE from XRT runtime incurred by new scheduler when
 invoking clEnqueueUnmapMemObject call, seems to be a bug. See Note (1).
* Altered the SYCL Library build runtime to optionally compile
  extensions when XRT is found. This is up for discussion. See Note (2).

Note (1): The new scheduler showcased this fun to find bug, it appears to be more
of an issue in XRT than the scheduler implementation however (I have
opened an issue for clarification here:
https://github.com/Xilinx/XRT/issues/1425). The scheduler seems to
handle generation of the dependent events fine and passes the correct
dependent events to the clEnqueueUnmapMemObject, however XRT doesn't
wait on the events to play out before invoking the check that errors
out. However, it seems that their should be a wait in place for the
prior events as this error call can depend on the events.. for example
in the case we have a clEnqueueMapBuffer call that isn't quite complete
yet.

Note (2): If XRT is found while building we define __SYCL_XILINX_ONLY__
this is a define that is currently shared with the compiler, it forces
certain sections of the code to be optionally built. These optional
sections contain XRT CL extensions as well as Intel sections. They are
chosen based on the vendor of the platform (e.g. if xilinx use xilinx
extensions). This is ONLY required in source (.cpp) files because they
are compiled to the libSYCL.so library, it can be excluded from .hpp
files. You can see this in effect in the current buffer_impl.hpp and memory_manager.cpp files.

Perhaps there is a better way to do this, one option may just be to have
the build system download and install XRT from Github. Similar to what
Intel seem to already to with the OpenCL ICD and headers.