[SYCL][Test][Runtime] reqd_work_group_size port and modification
reqd_work_group_size of 1,1,1 is now always applied to single_task.

It is now only applicable to parallel_for's (without an OpenCL runtime 
exception at least) with explicity specified local work group size via 
an ND_Range. With ONE exception, which is the case where a parallel_for 
has a global size of 1 and a dimension of 1, so undistinguishable from a 
single_task.

Altered reqd_work_group_size test to reflect the changes.