[SYCL][Runtime] Simplifying the spirv/spir builtin macro redirect approach and splitting into files
This change is toning down the more generic macro redirect from a prior
patch to something perhaps a little more easier to maintain for other
people for the time being. No need to overcomplicate it until we have to
really (e.g. more builtins). This also avoids difficult to reason about
upstream merges for us as the spirv builtins aren't mixed in with the
spir builtins.

Also seperating into spirv_vars/spir_var files that get optionally
included depending on the compile for device at the moment. This means
we shouldn't get as many merge conflicts when pulling upstream, but it
does mean we need to pay a little more attention to what happens in
spirv_vars as it needs to be reflected in spir_vars.