[SYCL][Driver] Make the Driver/KernelPropGen pass agnostic to the folder the file being compiled is in
Found this to be a problem when creating a run_tests.sh script that runs 
tests using a path outside of the directory e.g. 
xocc_tests/simple_tests/some_test.cpp.

In the case of KernelPropGen you'd end up with a weird name for the 
generated .sh that includes the directory in it, not ideal as the file 
just gets dumped into tmp at the moment. And it would look in an weird 
directory because of spurious /'s.

In the case of sycl-xocc it would place files into /tmp/ but look in the 
wrong place for them inside the script.

This tweak fixes this, but it needs some TLC at some point to make sure 
it makes sense for:

1) Files from anywhere in the system
2) Files resting in the same location as the compiler is invoked from
3) Multiple parallel invocations of the script