[SYCL][Test] Addition of a run_tests script for Xilinx FPGA
Not sure if we want this in the repository, but it's a handy script for 
automating test runs for now. Which works as a quick fix in the long 
term until we have the time and inclination to work out the changes 
required to integrate into the LLVM lit infrastructure.