[SYCL][Test] Adding two new tests that have bugs that need fixes
agdce_ice.cpp - is a test that will cause an ICE for xocc HLS when it 
tries to consume the LLVM/spir-df IR kernel, I passed this onto someone 
in the HLS team to have a look at for now.

boost_hana_functor_arg.cpp - seems to be address space inference 
related, an imposed address space cast seems to be breaking the inlining 
of some boost_hana functionallity. Needs some further investigation.