[SYCL][Driver] Add fsycl-allow-variadic-func argument
This allows variadic functions inside SYCL device code and is on by 
default for xocc compilation and off by default otherwise. It's a cc1 
driver option and requires -Xclang arg if you wish to specify it on the 
command line.

It's been added and is on by default for xocc as our ssdm compiler 
intriniscs are considered variadics, at least the way they're declared 
for now. This also allows the use of printf, which is sort of useful for 
us for sw_emu at the moment as the SYCL stream functionallity has yet to 
be tested.