[SYCL][sycl-xocc][opt] Clang/LLVM 9 byval workaround for xocc
This adds some tweaks to code in the XOCCIRDowngrader.cpp, remove 
ByVal's new syntanx and then readd old syntax, then change function 
block names to conform with Clang/LLVM 7.

We also now emit LLVM IR rather than bitcode from our final set of opt 
passes and then compile to bitcode using the Clang/LLVM 9 assembler 
packaged with SDX.

We do this because the ByVal attribute has went through some significant 
changes in upstream Clang that makes it impossible to output bitcode 
that conforms to older versions of Clang/LLVM without alterations to the 
Bitcode Writer (that would be an alternative fix for this problem, but 
let's try this one for the moment and see how hard it is to live with).

This can be removed as XOCC catches up to the patchs related to 
Phabricator Review: D62319