[SYCL] New Option sycl-header-only-library, notifies compiler no need for -lsycl
At the moment the compiler will automatically try and add -lsycl and 
search for a SYCL library, we won't use a library as ACAP++ aims to be 
header only. So this option is a requirement for seperating ACAP++ from 
the Intel SYCL runtime for now. 

In reality if you've compiled the compiler on the architecture you 
ineveitablly want to target, this won't matter at all you can let it 
include the ACAP++ headers and libsycl.so. This is a convienience option 
for cross-compilation for the most part. But also useful to make sure 
there is no eroneous code injected by including the Intel SYCL runtime.