[SYCL][Chess Scripts] Initial Chess compiler driver scripts and chess-clang wrapper
Initial 1 kernel/Elf binary compialtion script, it should offload and 
wrap the elf binary into the final kernel but there is no host side API 
for loading and unload this via program manager or the libxaiengine 
runtime for truly testing this yet. 

The kernel elf binary is currently dropped into the /tmp/ directory via 
a cp inside of the script. This is so we don't have to use any offloader 
related concepts and we can just load the elf to the tile by loading in 
from a seperate file. So host and device binary can be tested seperately 
essentially.