[SYCL][AIE][SemaSYCL] InitExpr Base Class Default Constructor Temp Fix
This seems to be only required if we wish to pass seperate arguments 
across rather than a full Tile class, so if it's possible to memcpy 
across a full Tile instance to a tile that may be a better option.

This is hopefully a temporary fix until Intel fix it upstream, it 
currently default constructs the base class which is fine for us at the 
moment. But this won't scale if we want non-default constructable base 
classes or to retain information. It's effectively slicing at the 
moment. To do it properly would require recursive functions checking 
base classes for parameters to pass through to the kernel function and 
also generation of the appropriate store/construction of the base 
classes.

At the moment if you're functor has a base class the load/stores will be 
off by the number of base classes so you won't have the correct number 
of stores and in cases where there are different types it will result in 
stores to the wrong Type