[SYCL][Driver][sycl-chess] Teach script to link multiple genearted mains
Scrpt will now invoke KernelPropGen invoke the generated bash script for 
the information it requires and then execute the required compilation 
and linking of all of the main files against the module currently being 
compiled. This is quite slow for the full 50x8 but the compile process 
appears to function. However, I would currently suggest only compiling 
for device for one_pe for the moment and then compiling the host 
seperately for the full board without the device compilation step as it 
takes a significant amount of time to compile everyting for device and 
only one binary is used in all cases at the moment..