[SYCL][Driver][OffloadWrapper][sycl-chess] First pass AIE Elf Kernel Offloader
This works and avoids phase alterations and it is fairly simple, it does 
require some modifications to the offload wrapper. I think even with 
phase alterations there is perhaps no way to avoid offload wrapper 
alterations. But I think this avenue is worth investigating at some 
point in the future.

The thing I am most irriated about in the offload wrapper changes is 
having to modify some of the StringRef members of the OffloadWrapper's 
Image class to String to keep some of our unpackaged arguments 
persistant across the wrapper. As otherwise they will fall out of scope 
at the end of the image unbundling function and there data is lost. This 
could be fixed by some refactoring in the future though I believe.