[SYCL][Test] Add Issue Relating To Local Memory
Currently Local Memory appears to be allocated as Global Memory, this 
appear to happen despite the correct address space qualification.

This requires further discussion with the HLS team, but I'm led to 
believe that passing a local pointer to the OpenCL kernel function may 
not be enough. HLS in it's C++ form represents local memory as a "stack" 
created array that will get allocated to BRAM.