[SYCL][Opt][sycll-xocc] Introduce InferAddressSpace pass and modify it to take an arguement for "FlatAddressSpace"
This pass will deduce generic address spaces to concrete address spaces, 
so DISABLE_INFER_AS=1 no longer needs to be defined in your enviornment.

There is still some issues with the edge_detection.cpp example as it 
unfortunately has some load/stores that are unoptimizeable for now it 
seems.. but the pass seems to work reasonably. This was 
introduced/forced my hand to deal with the commit: 
https://github.com/intel/llvm/pull/348/files as this pass disregards 
legal address space casts on the device and relies on them being deduced 
for SPIR devices/doesn't care and sticks them in generic for other 
devices. In particular in the Impl calls and Subscript operators, 
assignment of ConcretePtrType to either the PtrType or RefType forces an 
address space cast which isn't legal without generics. So an alternative 
fix/in addition to this if we encounter trouble, may be to reduce the 
amount of Address Space casts being made by reverting the referenced 
patch and redifining concrete address spaces for all PtrTypes for Xilinx 
or SPIR devices. 

In either case this pass should have been added a while ago, we'll see 
how it plays out.

We currently still use the -asfix pass, although it's probably not 
needed, I've yet to test the impact of having it/not having it/the 
placement in the pass list.

To add this pass and circumvent having to define a Target Infrastructure 
that we will barely use at the moment or add a sort of hacky use of this 
function transformation pass in InSPIRation I added a new option to 
InferAddressSpaces.cpp, which seems like the right thing to do.. and it 
may be upstreamable if there is any interest.