[SYCL][Opt] Add non-numeric argument renaming to non-kernel and non-intrinsic functions
This kills XOCC compilation at the moment, if you have a user defined 
functiton Clang/LLVM will now give it a numeric name where previously it 
would be unnamed. Interestingly XOCC (GCC in particular) seems to be 
fine with either no name or an alphabetic name, but not a numeric name.

Quick fix is to apply the same treatment as the Kernel and give them 
names from arg_0...arg_N.