[SYCL][sycl-chess] Merge duplicate functions
This uses the llvm merge function opt pass, and requires to run all opt
passes to be in specific order.
1. chess massage: reorder functions, change the linkage type, generate
     unmerged kernel names
2. merge func: merge and remove duplicate functions
3. kernel prop gen: revert the linkage type to original, generate merged
     kernel names

After these opt passes, this script uses the unmerged / merged kernel
names from passes to determine and skip the compilation for duplicated
kernels.

This works with certain assumptions between multiple passes, which is
described above and might be cleaned up later. Also it only optimizes
the compilation time, not output size. For the size optimization,
maybe the integration header should be used to notify the host
compliation to pick up the merge details. But the header is generated
in clang sema, before llvm merge pass, so it's not clear how it can
be done at the moment. The AIE binaries seem relative small part of
final binary, in terms of size anyway.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>