[SYCL][sycl-chess] simplify the merge handling
Instead of handling with additional scripting, re-use the image creation
routine that already exists. Use the fact that merged and unmerged
lists are identically ordered, and only the unmerged kernels are
compiled and have corresponding elf files. When creating an image, if
the elf file doesn't exist for given kernel, find and use the merged
kernel elf instead.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>