[SYCL][FPGA] Fix tests and make them more robust
Each xocc test is now compiled and executed in its own directory which
which is cleaned before starting the test to avoid pid conflict issue
that makes xsim prompt for user input.

These conflicts are now unavoidable due to the execution of the tests
in their own pid namespace, in order to clean the mess than hw emulation
can leave behind.