[VXX][SYCL] Add support for --vitis-ip-part during the SYCL compilation flow.
This has the effect of compiling the SYCL device code with the ip export flow which is much faster
than the V++ flow because it compiles only the kernel and not the kernel + the shell.

/!\ BUT the resulting executable WILL NOT WORK because it has not shell /!\

so this flow only exists as a very fast way to get performance reports for SYCL kernel code.